Module X509.Extension

X.509v3 extensions

type key_usage = [

| `Digital_signature

| `Content_commitment

| `Key_encipherment

| `Data_encipherment

| `Key_agreement

| `Key_cert_sign

| `CRL_sign

| `Encipher_only

| `Decipher_only

]

The polymorphic variant of key usages.

type extended_key_usage = [

| `Any

| `Server_auth

| `Client_auth

| `Code_signing

| `Email_protection

| `Ipsec_end

| `Ipsec_tunnel

| `Ipsec_user

| `Time_stamping

| `Ocsp_signing

| `Other of Asn.oid

]

The polymorphic variant of extended key usages.

type authority_key_id = Cstruct.t option * General_name.t * Z.t option

The authority key identifier, as present in the Authority Key Identifier extension.

type priv_key_usage_period = [

| `Interval of Ptime.t * Ptime.t

| `Not_after of Ptime.t

| `Not_before of Ptime.t

]

The private key usage period, as defined in RFC 3280.

type name_constraint = (General_name.b * int * int option) list

Name constraints, as defined in RFC 5280.

type policy = [

| `Any

| `Something of Asn.oid

]

Certificate policies, the policy extension.

type reason = [

| `Unspecified

| `Key_compromise

| `CA_compromise

| `Affiliation_changed

| `Superseded

| `Cessation_of_operation

| `Certificate_hold

| `Remove_from_CRL

| `Privilege_withdrawn

| `AA_compromise

]

Type of revocation reasons for a given distribution point.

type distribution_point_name = [

| `Full of General_name.t

| `Relative of Distinguished_name.t

]

Distribution point name, either a full one using general names, or a relative one using a distinguished name.

type distribution_point = distribution_point_name option * reason list option * General_name.t option

Distribution point, consisting of an optional name, an optional list of allowed reasons, and an optional issuer.

type 'a extension = bool * 'a

The type of an extension: the critical flag and the value itself.

type _ k =

| Unsupported : Asn.oid -> Cstruct.t extension k

| Subject_alt_name : General_name.t extension k

| Authority_key_id : authority_key_id extension k

| Subject_key_id : Cstruct.t extension k

| Issuer_alt_name : General_name.t extension k

| Key_usage : key_usage list extension k

| Ext_key_usage : extended_key_usage list extension k

| Basic_constraints : (bool * int option) extension k

| CRL_number : int extension k

| Delta_CRL_indicator : int extension k

| Priv_key_period : priv_key_usage_period extension k

| Name_constraints : (name_constraint * name_constraint) extension k

| CRL_distribution_points : distribution_point list extension k

| Issuing_distribution_point : (distribution_point_name option * bool * bool * reason list option * bool * bool) extension k

| Freshest_CRL : distribution_point list extension k

| Reason : reason extension k

| Invalidity_date : Ptime.t extension k

| Certificate_issuer : General_name.t extension k

| Policies : policy list extension k

The type of supported X509v3 and CRL extensions.

include Gmap.S with type 'a key = 'a k

type !'a0 key = 'a k

type t

val empty : t

val singleton : 'a key -> 'a -> t

val is_empty : t -> bool

val cardinal : t -> int

val mem : 'a key -> t -> bool

val find : 'a key -> t -> 'a option

val get : 'a key -> t -> 'a

val add_unless_bound : 'a key -> 'a -> t -> t option

val add : 'a key -> 'a -> t -> t

val remove : 'a key -> t -> t

val update : 'a key -> ('a option -> 'a option) -> t -> t

type b =

| B : 'a key * 'a -> b

val min_binding : t -> b option

val max_binding : t -> b option

val any_binding : t -> b option

val bindings : t -> b list

type eq = {

f : 'a. 'a key -> 'a -> 'a -> bool;

}

val equal : eq -> t -> t -> bool

type mapper = {

f : 'a. 'a key -> 'a -> 'a;

}

val map : mapper -> t -> t

val iter : (b -> unit) -> t -> unit

val fold : (b -> 'a -> 'a) -> t -> 'a -> 'a

val for_all : (b -> bool) -> t -> bool

val exists : (b -> bool) -> t -> bool

val filter : (b -> bool) -> t -> t

type merger = {

f : 'a. 'a key -> 'a option -> 'a option -> 'a option;

}

val merge : merger -> t -> t -> t

type unionee = {

f : 'a. 'a key -> 'a -> 'a -> 'a option;

}

val union : unionee -> t -> t -> t

val critical : 'a key -> 'a -> bool

critical ext_key ext_value is the critical bit in ext_value.

val pp : t Fmt.t

pp ppf ext_map pretty-prints the extension map.